fixed comment: created new module format-time.js
