+++ src/mock/task.js
@@ -58,18 +58,5 @@ const generateTasks = (count) => {
     .map(generateTask);
 };
 
-const castTimeFormat = (value) => {
-  return value < 10 ? `0${value}` : String(value);
-};
 
-const formatTime = (date) => {
-  const hours = castTimeFormat(date.getHours() % 12);
-  const minutes = castTimeFormat(date.getMinutes());
-
-  const interval = date.getHours() > 11 ? `pm` : `am`;
-
-  return `${hours}:${minutes} ${interval}`;
-};
-
-
-export {generateTask, generateTasks, formatTime};
+export {generateTask, generateTasks};
